Port SP1 modifications
https://github.com/wasmerio/wasmer/compare/69720a4374dbeb56fac98431be56c5823ed70f70...wakabat:official-wasmer:e065880e79728b4d15ffc9d4eccd5153d249de2c

+++ lib/compiler/src/engine/link.rs
@@ -213,9 +213,9 @@ fn apply_relocation(
         RelocationKind::Aarch64AdrPrelPgHi21 => unsafe {
             let (reloc_address, delta) = r.for_address(body, target_func_address as u64);
 
-            let delta = delta as isize;
+            let delta = delta as i64;
             assert!(
-                ((-1 << 32)..(1 << 32)).contains(&delta),
+                ((-1i64 << 32)..(1i64 << 32)).contains(&delta),
                 "can't generate page-relative relocation with ±4GB `adrp` instruction"
             );
